refactor: moved libraries initialization logic in index.js
